refactor(auth): relocate the cookie-domain filter to the write side (ADR-0033 PR 4.2)
`_auth/_browser_cookie_filter.py` was write-time policy misfiled under a
`browser_` name. Three of its five call sites are intent writers in
`storage.py` (`replace_from_remint`, `replace_from_login`,
`persist_minted_jar`), each of which had to reach the filter through a
function-local import from the leaf; only two are the browser-capture arms,
which were simply its first callers. Its content now lives in `storage.py`
beside the writers, and the leaf is a re-export shim.

Behaviour-identical by construction: same function objects, same call order,
same `notebooklm.auth` logger (both modules bound it by NAME, not `__name__`,
so no log-emission point moved and no warning-path test changed).

BOTH FILTER PASSES STAY — this is the point of the PR, not an oversight
(plan D3, maintainer-decided after investigation):

* The capture arms filter BEFORE `heal_captured_state`, so the in-memory
  PSIDTS heal's routing preflight and recovery-jar build see domain-filtered
  rows. Dropping that pass would hand the heal the raw browser capture:
  sibling-product cookies from the same session plus duplicate names across
  domain variants — the #2054 duplicate-name / domain-priority ambiguity class.
* The writer's pass is ADR-0029's entry-path-independent guarantee: it holds
  for every caller of the writer, including ones that never filtered.

The old prose framed the second pass as "idempotent with the call above" —
i.e. as redundancy, the exact reading that would get one of them deleted. Each
site now states its own distinct reason, in `storage.py`'s two writer
docstrings and at both capture-arm call sites.

Bookkeeping:
* `_auth/storage.py` ceiling 2183 -> 2416 under a fresh
  `# sanctioned merge (ADR-0033)` annotation naming the donor
  (`_auth/_browser_cookie_filter.py`, 214 -> 29, i.e. -185). The +48 gap
  between donor shrink and recipient growth is the D3 comment correction and
  the new labelled section header; the annotation records that explicitly.
  `browser_capture.py` stays at its measured 1251 pin.
* The shim joins `test_consolidation_shims_are_identity_reexports` (the gate
  the earlier merges established) rather than getting a third sibling gate.
* `browser_capture.py`'s `from . import storage` is hoisted to module scope:
  its stated justification ("keeps the writer stack off this module's import
  path") is void once the filter import creates that edge anyway. Still bound
  as the MODULE, so the writers stay late-bound and the patch seam is
  unchanged.
* ADR-0030 c-PR5's logger bullet is amended in place: the `notebooklm.auth`
  contract is restated as the constraint on any future relocation, since a
  module logging by `__name__` would silently reroute these warnings.
* `docs/architecture.md`'s `_auth` index and repo tree updated.

Gates: full suite 13663 passed / 64 skipped / 3 xfailed; `tests/_guardrails/`
1305 passed / 3 xfailed; whole-tree `ruff format --check` + `ruff check`
clean; `mypy src/notebooklm` clean.
